+++ source/marchingcubes/MarchingCubesPainter.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <memory>
-#include <queue>
 #include <unordered_map>
 
 #include <glm/vec3.hpp>
@@ -18,8 +17,6 @@
 
 #include "MarchingCubesPipeline.h"
 
-class Chunk;
-class ChunkRenderer;
 
 namespace globjects
 {
@@ -48,15 +45,6 @@ public:
     virtual void onInitialize() override;
     virtual void onPaint() override;
 
-	bool useMipMap() const;
-	void setUseMipMap(bool useMipMap);
-    
-protected:
-    void setupGrid();
-    void setupProjection();
-    void setupOpenGLState();
-
-
 protected:
     /* pipeline*/
     MarchingCubesPipeline m_pipeline;
@@ -67,15 +55,4 @@ protected:
     gloperate::AbstractPerspectiveProjectionCapability * m_projectionCapability;
     gloperate::AbstractCameraCapability * m_cameraCapability;
 
-    /* members */
-    globjects::ref_ptr<gloperate::AdaptiveGrid> m_grid;
-
-    std::unordered_map<glm::vec3, globjects::ref_ptr<Chunk>> m_chunks;
-    globjects::ref_ptr<ChunkRenderer> m_chunkRenderer;
-
-    std::queue<glm::vec3> m_chunkQueue;
-
-	/* parameters */
-	bool m_useMipMap;
-	bool m_useMipMapChanged;
 };
