+++ source/marchingcubes/MarchingCubesPipeline.h
@@ -1,5 +1,8 @@
 #pragma once
 
+#include <globjects/base/ref_ptr.h>
+#include <globjects/Texture.h>
+
 #include <gloperate/pipeline/AbstractPipeline.h>
 #include <gloperate/pipeline/Data.h>
 
@@ -27,6 +30,11 @@ public:
     gloperate::Data<gloperate::AbstractViewportCapability *> viewport;
     gloperate::Data<gloperate::AbstractCameraCapability *> camera;
     gloperate::Data<gloperate::AbstractPerspectiveProjectionCapability *> projection;
+
+    gloperate::Data<bool> useMipMap;
+
+    globjects::ref_ptr<globjects::Texture> colorTexture;
+    globjects::ref_ptr<globjects::Texture> groundTexture;
 
     gloperate::Data<gloperate::AbstractTargetFramebufferCapability *> targetFBO;
 };
