+++ source/marchingcubes/RenderStage.h
@@ -1,6 +1,11 @@
 #pragma once
 
+#include <queue>
+
+#include <glm/vec3.hpp>
+
 #include <globjects/base/ref_ptr.h>
+#include <globjects/Texture.h>
 
 
 #include <gloperate/pipeline/AbstractStage.h>
@@ -12,9 +17,15 @@ namespace gloperate
 {
     class AbstractViewportCapability;
     class AbstractVirtualTimeCapability;
-    class AbstractProjectionCapability;
+    class PerspectiveProjectionCapability;
     class AbstractCameraCapability;
+    class AbstractTargetFramebufferCapability;
+    class AdaptiveGrid;
 }
+
+
+class Chunk;
+class ChunkRenderer;
 
 
 class RenderStage : public gloperate::AbstractStage
@@ -27,7 +38,11 @@ public:
 public:
     gloperate::InputSlot<gloperate::AbstractViewportCapability *> viewport;
     gloperate::InputSlot<gloperate::AbstractCameraCapability *> camera;
-    gloperate::InputSlot<gloperate::AbstractProjectionCapability *> projection;
+    gloperate::InputSlot<gloperate::PerspectiveProjectionCapability *> projection;
+    gloperate::InputSlot<gloperate::AbstractTargetFramebufferCapability *> targetFBO;
+    gloperate::InputSlot<bool> useMipMap;
+    globjects::ref_ptr<globjects::Texture> colorTexture;
+    globjects::ref_ptr<globjects::Texture> groundTexture;
 
 
 protected:
@@ -35,9 +50,17 @@ protected:
 
     void render();
 
+    void setupGrid();
+    void setupProjection();
+    void setupOpenGLState();
+
 
 protected:
+    globjects::ref_ptr<gloperate::AdaptiveGrid> m_grid;
 
+    std::unordered_map<glm::vec3, globjects::ref_ptr<Chunk>> m_chunks;
+    globjects::ref_ptr<ChunkRenderer> m_chunkRenderer;
+    std::queue<glm::vec3> m_chunkQueue;
 
 };
